add edge case tests
